# src/main/resources/static/swagger.yaml
openapi: 3.0.3
info:
  title: API de Pagamentos
  description: |
    Esta é uma API para possibilitar o recebimento de pagamentos de débitos de pessoas físicas e jurídicas.
    Você pode encontrar mais informações sobre Swagger em [https://swagger.io](https://swagger.io).
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Servidor Local
tags:
  - name: pagamento
    description: Operações relacionadas a pagamentos
  - name: pessoa
    description: Operações relacionadas a pessoas
  - name: conta
    description: Operações relacionadas a contas
  - name: cartao
    description: Operações relacionadas a cartões
paths:
  /cadastro/pessoa:
    post:
      tags:
        - pessoa
      summary: Criar um novo cadastro de pessoa
      operationId: criarCadastroPessoa
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Pessoa'
      responses:
        '200':
          description: Cadastro de pessoa criado com sucesso
        '400':
          description: Erro ao criar cadastro de pessoa
  /cadastro/conta:
    post:
      tags:
        - conta
      summary: Criar uma nova conta associada a uma pessoa
      operationId: criarConta
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Conta'
      responses:
        '200':
          description: Conta criada com sucesso
        '400':
          description: Erro ao criar conta
  /cartao/credito:
    post:
      tags:
        - cartao
      summary: Criar um novo cartão de crédito associado a uma conta e pessoa
      operationId: criarCartaoCredito
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CartaoCredito'
      responses:
        '200':
          description: Cartão de crédito criado com sucesso
        '400':
          description: Erro ao criar cartão de crédito
  /cartao/debito:
    post:
      tags:
        - cartao
      summary: Criar um novo cartão de débito associado a uma conta e pessoa
      operationId: criarCartaoDebito
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CartaoDebito'
      responses:
        '200':
          description: Cartão de débito criado com sucesso
        '400':
          description: Erro ao criar cartão de débito
  /pagamento/realizar:
    post:
      tags:
        - pagamento
      summary: Realizar um novo pagamento
      operationId: realizarPagamento
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Pagamento'
      responses:
        '200':
          description: Pagamento realizado com sucesso
        '400':
          description: Erro ao realizar pagamento
  /pagamento/consultar:
    get:
      tags:
        - pagamento
      summary: Consultar pagamentos
      operationId: consultarPagamentos
      parameters:
        - name: codigoDebito
          in: query
          description: Código do débito
          schema:
            type: integer
            format: int64
        - name: cpfCnpj
          in: query
          description: CPF ou CNPJ do pagador
          schema:
            type: string
        - name: status
          in: query
          description: Status do pagamento
          schema:
            type: string
      responses:
        '200':
          description: Lista de pagamentos consultados com sucesso
        '400':
          description: Erro ao consultar pagamentos
  /pagamento/excluir:
    post:
      tags:
        - pagamento
      summary: Excluir um pagamento
      operationId: excluirPagamento
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PagamentoExclusao'
      responses:
        '200':
          description: Pagamento excluído com sucesso
        '400':
          description: Erro ao excluir pagamento
  /pagamento/alterar:
    post:
      tags:
        - pagamento
      summary: Alterar o status de um pagamento
      operationId: alterarStatusPagamento
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PagamentoAlteracao'
      responses:
        '200':
          description: Status do pagamento alterado com sucesso
        '400':
          description: Erro ao alterar status do pagamento
components:
  schemas:
    Pessoa:
      type: object
      properties:
        nome:
          type: string
        dataNascimento:
          type: string
          format: date
        genero:
          type: string
        endereco:
          type: string
        telefone:
          type: string
        email:
          type: string
        cpf:
          type: string
        rg:
          type: string
        estadoCivil:
          type: string
        nacionalidade:
          type: string
        nomePai:
          type: string
        nomeMae:
          type: string
        estado:
          type: string
        pais:
          type: string
    Conta:
      type: object
      properties:
        numeroConta:
          type: string
        saldo:
          type: number
        tipoConta:
          type: string
        cnpj:
          type: string
        titular:
          $ref: '#/components/schemas/Pessoa'
    CartaoCredito:
      type: object
      properties:
        numeroCartao:
          type: string
        nomeTitular:
          type: string
        dataValidade:
          type: string
        cvv:
          type: string
        limite:
          type: number
        conta:
          $ref: '#/components/schemas/Conta'
    CartaoDebito:
      type: object
      properties:
        numeroCartao:
          type: string
        nomeTitular:
          type: string
        dataValidade:
          type: string
        cvv:
          type: string
        conta:
          $ref: '#/components/schemas/Conta'
    Pagamento:
      type: object
      properties:
        codigoDebito:
          type: integer
          format: int64
        cpfCnpjPagador:
          type: string
        metodoPagamento:
          type: string
        numeroCartao:
          type: string
        valorPagamento:
          type: number
        status:
          type: string
    PagamentoExclusao:
      type: object
      properties:
        id:
          type: integer
          format: int64
    PagamentoAlteracao:
      type: object
      properties:
        id:
          type: integer
          format: int64
        status:
          type: string